Seção 9

+++ Back/src/ProEventos.API/Controllers/EventosController.cs
@@ -6,6 +6,7 @@ using ProEventos.Application.Contratos;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
 using System;
+using ProEventos.Application.Dtos;
 
 namespace ProEventos.API.Controllers
 {
@@ -26,8 +27,8 @@ namespace ProEventos.API.Controllers
         {
             try
             {
-                var eventos = await _eventoService.GetAllEventosByAsync(true);
-                if (eventos == null) return NotFound("Nenhum evento encontrado.");
+                var eventos = await _eventoService.GetAllEventosAsync(true);
+                if (eventos == null) return NoContent();
 
                 return Ok(eventos);
             }
@@ -44,7 +45,7 @@ namespace ProEventos.API.Controllers
             try
             {
                 var evento = await _eventoService.GetEventoByIdAsync(id, true);
-                if (evento == null) return NotFound("Evento por Id não encontrado.");
+                if (evento == null) return NoContent();
 
                 return Ok(evento);
             }
@@ -61,7 +62,7 @@ namespace ProEventos.API.Controllers
             try
             {
                 var evento = await _eventoService.GetAllEventosByTemaAsync(tema, true);
-                if (evento == null) return NotFound("Eventos por tema não encontrados");
+                if (evento == null) return NoContent();
 
                 return Ok(evento);
             }
@@ -73,7 +74,7 @@ namespace ProEventos.API.Controllers
         }
 
         [HttpPost]
-        public async Task<IActionResult> Post(Evento model)
+        public async Task<IActionResult> Post(EventoDto model)
         {
             try
             {
@@ -90,7 +91,7 @@ namespace ProEventos.API.Controllers
         }
 
         [HttpPut("{id}")]
-        public async Task<IActionResult> Put(int id, Evento model)
+        public async Task<IActionResult> Put(int id, EventoDto model)
         {
             try
             {
@@ -111,13 +112,17 @@ namespace ProEventos.API.Controllers
         {
            try
             {   
+
+                var evento = await _eventoService.GetEventoByIdAsync(id, true);
+                if (evento == null) return NoContent();
+
                 if(await _eventoService.DeleteEvento(id))
                 {
                     return Ok("Deletado");
                 }
                 else
                 {
-                    return BadRequest("Evento não deletado.");
+                    throw new Exception("Ocorreu um problema não específico ao tentar deletar Evento.");
                 }
             }
             catch (Exception ex)
